Move db models to db folder
