+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode
 *.sh
 !uptime.sh
+*.idea
