+++ Server/db/mongo/modules/checkModule.js
@@ -1,6 +1,6 @@
-const Check = require("../../../models/Check");
-const Monitor = require("../../../models/Monitor");
-const User = require("../../../models/User");
+const Check = require("../../models/Check");
+const Monitor = require("../../models/Monitor");
+const User = require("../../models/User");
 const logger = require("../../../utils/logger");
 const SERVICE_NAME = "checkModule";
 const dateRangeLookup = {
