+++ Server/db/mongo/modules/userModule.js
@@ -1,5 +1,5 @@
-const UserModel = require("../../../models/User");
-const TeamModel = require("../../../models/Team");
+const UserModel = require("../../models/User");
+const TeamModel = require("../../models/Team");
 const { errorMessages } = require("../../../utils/messages");
 const { GenerateAvatarImage } = require("../../../utils/imageProcessing");
