+++ Server/routes/maintenanceWindowRoute.js
@@ -1,7 +1,7 @@
 const router = require("express").Router();
 const maintenanceWindowController = require("../controllers/maintenanceWindowController");
 const { verifyOwnership } = require("../middleware/verifyOwnership");
-const Monitor = require("../models/Monitor");
+const Monitor = require("../db/models/Monitor");
 
 router.get(
   "/monitor/:monitorId",
