+++ Server/routes/settingsRoute.js
@@ -1,7 +1,7 @@
 const router = require("express").Router();
 const settingsController = require("../controllers/settingsController");
 const { isAllowed } = require("../middleware/isAllowed");
-const Monitor = require("../models/Monitor");
+const Monitor = require("../db/models/Monitor");
 
 router.get("/", settingsController.getAppSettings);
 router.put(
